requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.24.0
scipy>=1.10.0
opencv-python>=4.7.0
streamlit>=1.30.0
diffusers>=0.23.0
transformers>=4.35.2
accelerate>=0.20.0